Persist redux state
Previously, if the user reloaded the page or closed the browser, all state was lost. That means the user had to login again every time they came back to the app.

Now the state is persisted and the user sees whatever he was seing before leaving the page.

I refactored the authentication saga because after state was restored, it prevented the user from logging out. That happened because the saga expected the login action to come before the logout action. However, in the situation that the user is already authenticated then restored the persisted state, the log out action would come before.

Now the saga expects the two actions in parallel. Both username and password need to be save in order for Basic Authentication to be restored.