Fix restoration of auth credentials
The dashboard was unable to fetch new projects on page reload because the credentials were not being restored before the request.

That happened because the restoration was happening in componentDidMount. By the time componentDidMount was called, the dashboard had already started the request without credentials.

I moved the code into the constructor because it executes before any other child component.